src/siaa_vault/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
aiosqlite==0.20.0
pydantic==2.9.2
pydantic-settings==2.5.2
cryptography==43.0.1
PyJWT==2.9.0
httpx==0.27.2
apscheduler==3.10.4
loguru==0.7.2
